fix: correct import utility placement boundary
Separate provider-agnostic utilities from DnD Beyond-specific utilities:
- Move ModifierLike interface, isPresent, escapeRegExp, ABILITY_KEYS to utils.ts
- Move flattenModifiers from monolith to dndBeyond-utils.ts
- Remove DndBeyondModifier import from utils.ts (now uses ModifierLike)
- Update all import paths affected by these moves

Fixes the architectural boundary so utils.ts has no provider-specific
imports and dndBeyond-utils.ts contains only DnD Beyond-specific code.

Closes #173 #176 #159
